move components to own directory
